shacl_validator/shacl_validator_grpc_py/pyproject.toml
# Copyright 2025 Lincoln Institute of Land Policy
# SPDX-License-Identifier: Apache-2.0

[project]
name = "shacl-validator-grpc-py"
version = "0.1.0"
description = "A grpc service for validating shacl shapes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "pyshacl>=0.30.1",
    "rdflib>=7.1.4",
    "grpcio-tools>=1.74.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "ruff>=0.12.10",
]


[tool.pyright]
# ignore generated code
ignore = ["*pb2_grpc.py"]


[tool.ruff]
exclude = [
    "*pb2_grpc.py"
]